+++ app/src/main/AndroidManifest.xml
@@ -11,6 +11,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.RickMortyChars"
-        android:usesCleartextTraffic="true"></application>
+        android:usesCleartextTraffic="true" />
 
 </manifest>
