+++ build.gradle
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     dependencies {
-        classpath "com.google.dagger:hilt-android-gradle-plugin:2.50"
+        classpath libs.hilt.android.gradle.plugin
     }
     ext {
         versions = [
