+++ data/build.gradle
@@ -1,8 +1,6 @@
 plugins {
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
-    id 'com.google.devtools.ksp'
-    id 'dagger.hilt.android.plugin'
 }
 
 android {
@@ -34,41 +32,6 @@ android {
 
 dependencies {
     implementation libs.androidx.core
-    implementation project(path: ':domain:dataapi')
-
-    // room
-    implementation libs.room.runtime
-    implementation libs.room
-    ksp libs.room.compiler
-
-    // coroutines, flow
-    implementation libs.coroutines.core
-    implementation libs.coroutines.android
-
-    // json
-    implementation libs.moshi.adapters
-    implementation libs.moshi.kotlin
-    implementation libs.moshi.kotlin.codegen
-    implementation libs.moshi
-
-    // di
-    implementation libs.hilt.android
-    ksp libs.hilt.android.compiler
-
-    // net
-    implementation libs.okhttp
-    implementation libs.okhttp.logging.interceptor
-
-    // rest
-    implementation libs.retrofit.converter.moshi
-    implementation libs.retrofit
-
-    // paging
-    implementation libs.paging.runtime
-    implementation libs.paging.room
-
-    // logging
-    implementation libs.timber
 
     // testing
     testImplementation libs.junit
