+++ data/common/src/main/java/io/chthonic/rickmortychars/data/common/CommonSingletonModule.kt
@@ -1,0 +1,31 @@
+package io.chthonic.rickmortychars.data.common
+
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal class CommonSingletonModule {
+    @Provides
+    @Singleton
+    fun provideMoshi(): Moshi = Moshi.Builder()
+        .addLast(KotlinJsonAdapterFactory())
+        .build()
+
+    @Provides
+    @Singleton
+    fun provideOkHttpClient(): OkHttpClient = OkHttpClient.Builder()
+        .addInterceptor(
+            HttpLoggingInterceptor(
+                logger = HttpLoggingInterceptor.Logger.DEFAULT
+            ).apply {
+                level = HttpLoggingInterceptor.Level.BODY
+            }).build()
+}
