+++ data/rickmorty/src/main/java/io/chthonic/rickmortychars/data/rickmorty/RickMortySingletonModule.kt
@@ -1,19 +1,17 @@
-package io.chthonic.rickmortychars.data
+package io.chthonic.rickmortychars.data.rickmorty
 
 import android.content.Context
 import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import io.chthonic.rickmortychars.data.rest.RickMortyRestApi
-import io.chthonic.rickmortychars.data.database.CharactersDao
-import io.chthonic.rickmortychars.data.database.DatabaseFactory
-import io.chthonic.rickmortychars.data.database.RickMortyDatabase
+import io.chthonic.rickmortychars.data.rickmorty.rest.RickMortyRestApi
+import io.chthonic.rickmortychars.data.rickmorty.database.CharactersDao
+import io.chthonic.rickmortychars.data.rickmorty.database.DatabaseFactory
+import io.chthonic.rickmortychars.data.rickmorty.database.RickMortyDatabase
 import io.chthonic.rickmortychars.domain.dataapi.RickMortyRepository
 import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import javax.inject.Singleton
@@ -22,22 +20,7 @@ private const val BASE_URL = "https://rickandmortyapi.com/api/"
 
 @Module
 @InstallIn(SingletonComponent::class)
-internal class DataSingletonModule {
-    @Provides
-    @Singleton
-    fun provideMoshi(): Moshi = Moshi.Builder()
-        .addLast(KotlinJsonAdapterFactory())
-        .build()
-
-    @Provides
-    @Singleton
-    fun provideOkHttpClient(): OkHttpClient = OkHttpClient.Builder()
-        .addInterceptor(
-            HttpLoggingInterceptor(
-                logger = HttpLoggingInterceptor.Logger.DEFAULT
-            ).apply {
-                level = HttpLoggingInterceptor.Level.BODY
-            }).build()
+internal class RickMortySingletonModule {
 
     @Provides
     @Singleton
