+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-compose = '1.6.3'
+compose = '1.6.4'
 navigation = '2.7.7'
 lifecycle = '2.7.0'
 moshi = '1.15.1'
@@ -15,6 +15,12 @@ coil = '2.5.0'
 paging = '3.2.1'
 room-paging = '2.6.1'
 retrofit = '2.9.0'
+agp = "8.2.2"
+kotlin = "1.9.22"
+androidx-test-ext-junit = "1.1.5"
+espresso-core = "3.5.1"
+appcompat = "1.6.1"
+material = "1.11.0"
 
 [libraries]
 # android
@@ -23,6 +29,7 @@ androidx-core = { module = "androidx.core:core-ktx", version.ref = "androidx-cor
 # dp
 hilt-android = { module = "com.google.dagger:hilt-android", version.ref = "hilt" }
 hilt-android-compiler = { module = "com.google.dagger:hilt-android-compiler", version.ref = "hilt" }
+hilt-android-gradle-plugin = { module = "com.google.dagger:hilt-android-gradle-plugin", version.ref = "hilt" }
 hilt-navigation_compose = { module = "androidx.hilt:hilt-navigation-compose", version.ref = "hilt_compose" }
 
 # loggin
@@ -77,3 +84,13 @@ coil-compose = { module = "io.coil-kt:coil-compose", version.ref = "coil" }
 
 # testing
 junit = { module = "junit:junit", version.ref = "junit" }
+androidx-test-ext-junit = { group = "androidx.test.ext", name = "junit", version.ref = "androidx-test-ext-junit" }
+espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espresso-core" }
+
+# ui
+appcompat = { group = "androidx.appcompat", name = "appcompat", version.ref = "appcompat" }
+material = { group = "com.google.android.material", name = "material", version.ref = "material" }
+
+[plugins]
+androidLibrary = { id = "com.android.library", version.ref = "agp" }
+kotlinAndroid = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
