+++ settings.gradle
@@ -19,3 +19,5 @@ include ':domain'
 include ':domain:dataapi'
 include ':domain:presentationapi'
 include ':presentation'
+include ':data:common'
+include ':data:rickmorty'
